#211392: Prevent from submit process waiting on flush tasks, which is causing deadlocks. A test is written to verify the correct functionality.

+++ uihandler/src/org/netbeans/modules/uihandler/Installer.java
@@ -179,6 +179,9 @@ public class Installer extends ModuleInstall implements Runnable {
     /** Flag to store status of last metrics upload */
     private static boolean logMetricsUploadFailed = false;
     
+    /** Log records currently displaying/uploading */
+    private static final ThreadLocal<List<LogRecord>> logRecords = new ThreadLocal<List<LogRecord>>();
+    
     private static final  String USAGE_STATISTICS_ENABLED          = "usageStatisticsEnabled"; // NOI18N
     private static final String USAGE_STATISTICS_SET_BY_IDE       = "usageStatisticsSetByIde"; // NOI18N
     private static final String USAGE_STATISTICS_NB_OF_IDE_STARTS = "usageStatisticsNbOfIdeStarts"; // NOI18N
@@ -315,6 +318,11 @@ public class Installer extends ModuleInstall implements Runnable {
         }
     }
     
+    /** Accessed from tests. */
+    static int getLogsSizeTest() {
+        return logsSize;
+    }
+    
     private void logIdeStartup() {
         Logger.getLogger("org.netbeans.ui").log(new LogRecord(Level.CONFIG, IDE_STARTUP));
     }
@@ -436,11 +444,22 @@ public class Installer extends ModuleInstall implements Runnable {
         closeLogStream();
         closeLogStreamMetrics();
     }
+    
+    static boolean isImmediateWriteOut(LogRecord r) {
+        List<LogRecord> preferredLog = logRecords.get();
+        if (preferredLog != null) {
+            preferredLog.add(r);
+            return true;
+        } else {
+            return false;
+        }
+    }
 
     static void writeOut(LogRecord r) {
         try {
             boolean logOverflow;
             boolean logSizeControl;
+            List<LogRecord> logs = null;
             synchronized (UIGESTURE_LOG_LOCK) {
                 LogRecords.write(logStream(), r);
                 logsSize++;
@@ -454,15 +473,19 @@ public class Installer extends ModuleInstall implements Runnable {
                 }
                 if (logOverflow) {
                     closeLogStream();
+                    if (isHintsMode()) {
+                        logs = new ArrayList<LogRecord>(getLogs());
+                    }
                 }
                 logSizeControl = (logsSize % 100) == 0 && !logOverflow;
             }
             if (logOverflow) {
                 if (isHintsMode()) {
+                    final List<LogRecord> recs = logs;
                     class Auto implements Runnable {
                         @Override
                         public void run() {
-                            displaySummary("WELCOME_URL", true, true,true);
+                            displaySummary("WELCOME_URL", true, true,true, DataType.DATA_UIGESTURE, recs, null);
                         }
                     }
                     RP.post(new Auto()).waitFinished();
@@ -588,10 +611,11 @@ public class Installer extends ModuleInstall implements Runnable {
             }
             if (upload) {
                 //Task to upload metrics data
+                final List<LogRecord> recs = null;
                 class Auto implements Runnable {
                     @Override
                     public void run() {
-                        displaySummary("METRICS_URL", true, true, true, DataType.DATA_METRICS, null);
+                        displaySummary("METRICS_URL", true, true, true, DataType.DATA_METRICS, recs, null);
                     }
                 }
                 //Must be performed out of lock because it calls getLogsMetrics
@@ -1021,7 +1045,7 @@ public class Installer extends ModuleInstall implements Runnable {
     
     private static AtomicReference<String> DISPLAYING = new AtomicReference<String>();
 
-    private static boolean displaySummary(String msg, boolean explicit, boolean auto, boolean connectDialog, DataType dataType, SlownessData slownData) {
+    private static boolean displaySummary(String msg, boolean explicit, boolean auto, boolean connectDialog, DataType dataType, List<LogRecord> recs, SlownessData slownData) {
         if (!DISPLAYING.compareAndSet(null, msg)) {
             return true;
         }
@@ -1036,7 +1060,7 @@ public class Installer extends ModuleInstall implements Runnable {
                 }
             }
 
-            Submit submit = auto ? new SubmitAutomatic(msg, Button.SUBMIT, dataType) : new SubmitInteractive(msg, connectDialog, dataType, slownData);
+            Submit submit = auto ? new SubmitAutomatic(msg, Button.SUBMIT, dataType, recs) : new SubmitInteractive(msg, connectDialog, dataType, recs, slownData);
             submit.doShow(dataType);
             v = submit.okToExit;
         } finally {
@@ -1046,11 +1070,11 @@ public class Installer extends ModuleInstall implements Runnable {
     }
 
     public static boolean displaySummary(String msg, boolean explicit, boolean auto, boolean connectDialog) {
-        return displaySummary(msg, explicit, auto, connectDialog, DataType.DATA_UIGESTURE, null);
+        return displaySummary(msg, explicit, auto, connectDialog, DataType.DATA_UIGESTURE, null, null);
     }
 
     static boolean displaySummary(String msg, boolean explicit, boolean auto, boolean connectDialog, SlownessData slownessData) {
-        return displaySummary(msg, explicit, auto, connectDialog, DataType.DATA_UIGESTURE, slownessData);
+        return displaySummary(msg, explicit, auto, connectDialog, DataType.DATA_UIGESTURE, null, slownessData);
     }
 
     /** used only in tests - low performance - should use read logs*/
@@ -1203,27 +1227,41 @@ public class Installer extends ModuleInstall implements Runnable {
             h.progress(NbBundle.getMessage(Installer.class, "MSG_UploadConnecting")); // NOI18N
         }
         
+        boolean fileProtocol = "file".equals(postURL.getProtocol()); // For tests or other non-http usages
+        
         LOG.log(Level.FINE, "uploadLogs, postURL = {0}", postURL); // NOI18N
-        URLConnection conn = postURL.openConnection();
+        URLConnection conn;
+        if (fileProtocol) {
+            conn = null;
+        } else {
+            conn = postURL.openConnection();
+        }
 
         if (dataType != DataType.DATA_METRICS) {
             h.progress(10);
         }
         
-        conn.setReadTimeout(60000);
-        conn.setDoOutput(true);
-        conn.setDoInput(true);
-        conn.setRequestProperty("Content-Type", "multipart/form-data; boundary=--------konec<>bloku");
-        conn.setRequestProperty("Pragma", "no-cache");
-        conn.setRequestProperty("Cache-control", "no-cache");
-        conn.setRequestProperty("User-Agent", "NetBeans");
+        if (!fileProtocol) {
+            conn.setReadTimeout(60000);
+            conn.setDoOutput(true);
+            conn.setDoInput(true);
+            conn.setRequestProperty("Content-Type", "multipart/form-data; boundary=--------konec<>bloku");
+            conn.setRequestProperty("Pragma", "no-cache");
+            conn.setRequestProperty("Cache-control", "no-cache");
+            conn.setRequestProperty("User-Agent", "NetBeans");
+        }
         
         if (dataType != DataType.DATA_METRICS) {
             h.progress(NbBundle.getMessage(Installer.class, "MSG_UploadSending"), 20);
         }
         LOG.log(Level.FINE, "uploadLogs, header sent"); // NOI18N
 
-        PrintStream os = new PrintStream(conn.getOutputStream());
+        PrintStream os;
+        if (fileProtocol) {
+            os = new PrintStream(new FileOutputStream(postURL.getFile()));
+        } else {
+            os = new PrintStream(conn.getOutputStream());
+        }
         /*
         os.println("POST " + postURL.getPath() + " HTTP/1.1");
         os.println("Pragma: no-cache");
@@ -1334,27 +1372,33 @@ public class Installer extends ModuleInstall implements Runnable {
         if (dataType != DataType.DATA_METRICS) {
             h.progress(NbBundle.getMessage(Installer.class, "MSG_UploadReading"), cnt + 10);
         }
+
+        //System.err.println("DONE: Uploaded logs to "+postURL+" in "+Thread.currentThread());
         
-        LOG.log(Level.FINE, "uploadLogs, reading reply"); // NOI18N
-        InputStream is = conn.getInputStream();
-        StringBuffer redir = new StringBuffer();
-        for (;;) {
-            int ch = is.read();
-            if (ch == -1) {
-                break;
+        StringBuffer redir = null;
+        Matcher m = null;
+        if (!fileProtocol) {
+            LOG.log(Level.FINE, "uploadLogs, reading reply"); // NOI18N
+            InputStream is = conn.getInputStream();
+            redir = new StringBuffer();
+            for (;;) {
+                int ch = is.read();
+                if (ch == -1) {
+                    break;
+                }
+                redir.append((char)ch);
             }
-            redir.append((char)ch);
-        }
-        is.close();
-        
-        if (dataType != DataType.DATA_METRICS) {
-            h.progress(cnt + 20);
-        }
+            is.close();
 
-        LOG.log(Level.FINE, "uploadLogs, Reply from uploadLogs: {0}", redir);
+            if (dataType != DataType.DATA_METRICS) {
+                h.progress(cnt + 20);
+            }
 
-        Pattern p = Pattern.compile("<meta\\s*http-equiv=.Refresh.\\s*content.*url=['\"]?([^'\" ]*)\\s*['\"]", Pattern.CASE_INSENSITIVE | Pattern.MULTILINE | Pattern.DOTALL);
-        Matcher m = p.matcher(redir);
+            LOG.log(Level.FINE, "uploadLogs, Reply from uploadLogs: {0}", redir);
+
+            Pattern p = Pattern.compile("<meta\\s*http-equiv=.Refresh.\\s*content.*url=['\"]?([^'\" ]*)\\s*['\"]", Pattern.CASE_INSENSITIVE | Pattern.MULTILINE | Pattern.DOTALL);
+            m = p.matcher(redir);
+        }
 
         if (isOOM){
             FileObject fo = FileUtil.createData(getHeapDump());
@@ -1369,17 +1413,21 @@ public class Installer extends ModuleInstall implements Runnable {
             lock.releaseLock();
         }
 
-        if (m.find()) {
-            LOG.log(Level.FINE, "uploadLogs, found url = {0}", m.group(1)); // NOI18N
-            return new URL(m.group(1));
+        if (!fileProtocol) {
+            if (m.find()) {
+                LOG.log(Level.FINE, "uploadLogs, found url = {0}", m.group(1)); // NOI18N
+                return new URL(m.group(1));
+            } else {
+                File f = File.createTempFile("uipage", "html");
+                f.deleteOnExit();
+                FileWriter w = new FileWriter(f);
+                w.write(redir.toString());
+                w.close();
+                LOG.log(Level.FINE, "uploadLogs, temporary url = {0}", f.toURI()); // NOI18N
+                return f.toURI().toURL();
+            }
         } else {
-            File f = File.createTempFile("uipage", "html");
-            f.deleteOnExit();
-            FileWriter w = new FileWriter(f);
-            w.write(redir.toString());
-            w.close();
-            LOG.log(Level.FINE, "uploadLogs, temporary url = {0}", f.toURI()); // NOI18N
-            return f.toURI().toURL();
+            return null;
         }
     }
 
@@ -1519,17 +1567,21 @@ public class Installer extends ModuleInstall implements Runnable {
         protected ExceptionsSettings settings;
         
         public Submit(String msg) {
-            this(msg,DataType.DATA_UIGESTURE);
+            this(msg,DataType.DATA_UIGESTURE, null);
         }
 
-        public Submit(String msg, DataType dataType) {
+        public Submit(String msg, DataType dataType, List<LogRecord> recs) {
             this.msg = msg;
             this.dataType = dataType;
             isSubmiting = new AtomicBoolean(false);
-            if (dataType == DataType.DATA_METRICS) {
-                recs = getLogsMetrics();
-            }else{
-                recs = new ArrayList<LogRecord>(getLogs());
+            if (recs != null) {
+                this.recs = recs;
+            } else {
+                if (dataType == DataType.DATA_METRICS) {
+                    this.recs = getLogsMetrics();
+                } else {
+                    this.recs = new ArrayList<LogRecord>(getLogs());
+                }
             }
             if ("ERROR_URL".equals(msg)) { // NOI18N
                 report = true;
@@ -1560,8 +1612,15 @@ public class Installer extends ModuleInstall implements Runnable {
         }
 
         public void doShow(DataType dataType) {
+            //System.err.println("doShow("+dataType+")");
+            //Thread.dumpStack();
             if (dataType == DataType.DATA_UIGESTURE) {
-                logDeactivated();
+                try {
+                    logRecords.set(recs);
+                    logDeactivated();
+                } finally {
+                    logRecords.remove();
+                }
             }
             findDD();
             
@@ -1745,6 +1804,8 @@ public class Installer extends ModuleInstall implements Runnable {
             String actionURL = decodeButtons(e.getSource(), universalResourceLocator, dataType);
 
             LOG.log(Level.FINE, "actionPerformed: command = {0}", e.getActionCommand()); // NOI18N
+            //System.err.println("actionPerformed: command = "+e.getActionCommand()); // NOI18N
+            //Thread.dumpStack();
 
             boolean submit = Button.SUBMIT.isCommand(actionURL);
             if (Button.AUTO_SUBMIT.isCommand(e.getActionCommand())) {
@@ -2020,11 +2081,11 @@ public class Installer extends ModuleInstall implements Runnable {
         }
 
         public SubmitInteractive(String msg, boolean connectDialog, DataType dataType) {
-            this(msg, connectDialog, dataType, null);
+            this(msg, connectDialog, dataType, null, null);
         }
 
-        private SubmitInteractive(String msg, boolean connectDialog, DataType dataType, SlownessData slownData) {
-            super(msg, dataType);
+        private SubmitInteractive(String msg, boolean connectDialog, DataType dataType, List<LogRecord> recs, SlownessData slownData) {
+            super(msg, dataType, recs);
             this.connectDialog = connectDialog;
             this.slownData = slownData;
         }
@@ -2417,13 +2478,13 @@ public class Installer extends ModuleInstall implements Runnable {
         final Button def;
         private boolean urlComputed;
 
-        public SubmitAutomatic(String msg, Button def, DataType dataType) {
-            super(msg,dataType);
+        public SubmitAutomatic(String msg, Button def, DataType dataType, List<LogRecord> recs) {
+            super(msg, dataType, recs);
             this.def = def;
         }
 
         public SubmitAutomatic(String msg, Button def) {
-            this(msg,def,DataType.DATA_UIGESTURE);
+            this(msg, def, DataType.DATA_UIGESTURE, null);
         }
 
         @Override
@@ -2453,6 +2514,8 @@ public class Installer extends ModuleInstall implements Runnable {
         }
         @Override
         protected Object showDialogAndGetValue(final DialogDescriptor dd) {
+            //System.err.println("showDialogAndGetValue()");
+            //Thread.dumpStack();
             while (!urlComputed) {
                 synchronized (this) {
                     try {
